Add radial menu lib; bump deps, gradle to latest
Since I'm going to be modifying the library (to place the menu at
absolute coordinates), but I don't want to deal with distribution, the
easiest way to get this proof of concept working is just to copy the
files into this project.

However, I had a hell of a time getting things to build when I did that.
In the end, I just bumped everything to the latest version, including
gradle. It's totally ugly, especially to review, but this isn't intended
to be merged anyway, so it's okay.